+++ klant/src/main/java/be/one16/barka/klant/domain/Werkuur.java
@@ -4,8 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.util.Date;
 import java.util.UUID;
 
 @Data
@@ -16,11 +16,11 @@ public class Werkuur {
     private UUID werkuurId;
     private LocalDate datum;
     private double aantalUren;
-    private double uurTarief;
+    private BigDecimal uurTarief;
     private int btwPerc;
-    private double totaalExclusBtw;
-    private double totaalInclusBtw;
-    private double btwBedrag;
+    private BigDecimal totaalExclusBtw;
+    private BigDecimal totaalInclusBtw;
+    private BigDecimal btwBedrag;
     private UUID orderId;
 
 }
